Moved matrix/random package to utils/random
