+++ core/src/main/java/hivemall/knn/similarity/DIMSUMMapperUDTF.java
@@ -22,10 +22,10 @@ import hivemall.UDTFWithOptions;
 import hivemall.fm.Feature;
 import hivemall.fm.IntFeature;
 import hivemall.fm.StringFeature;
-import hivemall.math.random.PRNG;
-import hivemall.math.random.RandomNumberGeneratorFactory;
 import hivemall.utils.hadoop.HiveUtils;
 import hivemall.utils.lang.Primitives;
+import hivemall.utils.random.PRNG;
+import hivemall.utils.random.RandomNumberGeneratorFactory;
 
 import java.util.ArrayList;
 import java.util.Arrays;
