+++ core/src/main/java/hivemall/smile/utils/SmileExtUtils.java
@@ -22,8 +22,6 @@ import hivemall.annotations.VisibleForTesting;
 import hivemall.math.matrix.ColumnMajorMatrix;
 import hivemall.math.matrix.Matrix;
 import hivemall.math.matrix.MatrixUtils;
-import hivemall.math.random.PRNG;
-import hivemall.math.random.RandomNumberGeneratorFactory;
 import hivemall.math.vector.VectorProcedure;
 import hivemall.smile.classification.DecisionTree.SplitRule;
 import hivemall.utils.collections.arrays.SparseIntArray;
@@ -32,6 +30,8 @@ import hivemall.utils.collections.lists.IntArrayList;
 import hivemall.utils.lang.NumberUtils;
 import hivemall.utils.lang.Preconditions;
 import hivemall.utils.math.MathUtils;
+import hivemall.utils.random.PRNG;
+import hivemall.utils.random.RandomNumberGeneratorFactory;
 import smile.data.NominalAttribute;
 import smile.data.NumericAttribute;
 import smile.sort.QuickSort;
