+++ core/src/test/java/hivemall/smile/classification/DecisionTreeTest.java
@@ -23,8 +23,6 @@ import static org.junit.Assert.assertEquals;
 import hivemall.math.matrix.Matrix;
 import hivemall.math.matrix.builders.CSRMatrixBuilder;
 import hivemall.math.matrix.dense.RowMajorDenseMatrix2d;
-import hivemall.math.random.PRNG;
-import hivemall.math.random.RandomNumberGeneratorFactory;
 import hivemall.math.vector.DenseVector;
 import hivemall.smile.classification.DecisionTree.Node;
 import hivemall.smile.classification.DecisionTree.SplitRule;
@@ -35,6 +33,8 @@ import hivemall.utils.codec.Base91;
 import hivemall.utils.lang.ArrayUtils;
 import hivemall.utils.lang.StringUtils;
 import hivemall.utils.math.MathUtils;
+import hivemall.utils.random.PRNG;
+import hivemall.utils.random.RandomNumberGeneratorFactory;
 import smile.data.Attribute;
 import smile.data.AttributeDataset;
 import smile.data.NominalAttribute;
